Fix for gcc 10 type conversion
Converted to byte first to remove error:

error: conversion from 'int' to non-scalar type
'std::variant<unsigned char>' requested |   299 |
std::variant<uint8_t> value = 0;

Tested: built with dunfell

+++ src/smbiosmdrv2handler.cpp
@@ -296,7 +296,7 @@ ipmi::RspType<std::vector<uint8_t>> mdr2GetDir(uint16_t agentId,
         return ipmi::responseParmOutOfRange();
     }
 
-    std::variant<uint8_t> value = 0;
+    std::variant<uint8_t> value = static_cast<uint8_t>(0);
     if (0 != mdrv2->sdplusMdrv2GetProperty("DirectoryEntries", value, service))
     {
         phosphor::logging::log<phosphor::logging::level::ERR>(
